Change how letter closes if not specified
The letter will default to `Yours faithfully,` because opening defaults
to `Dear Sir or Madam`. If opening is specified (assuming this is
because you know the name of the person you are writing to) closing will
default to `Yours Sincerely`. This behaviour can still be overidden by
specifying `closing: ...` in the yaml header