VIM: Wiping vim configuration
My old vim config has a lot in there that I no longer use and as a
result I will be starting again. My vim configuration used to be a
submodule of my dotfiles, this proved to be more hassel than it was
worth so from now my new vim configuration will be in my main dotfiles
repository in the vim/ directory.

I will be re-building my vim configuration and plan do document
everything as I go in commit messages as seen in Greg Hurrel's excelent
dotfiles repo: https://github.com/wincent/wincent

I will try to keep the number of non-syntax plugins to a minimum.