VIM: Adds autoload function for writing mode
Writing mode is to be used when I am writing prose rather than code, for
example markdown files or emails.

Currently, it enables spellcheck and adds insert mode mappings that add
undo points when certain punctuation is inserted, such as full stops or
commas.

Currently this is used for markdown files, emails and git commits