SHELL: Adds aliases for cp and mv with progress bars
I found patches at the link below for mv and cp that add options for
progress bars. I have applied and compiled them and named the commands
cpg and mvg as recommended in the readme. They are named like that (I
assume) because they add a -g flag for progress.

If they are present and in my path, I always want to use them with the
progress bar flag in place of cp or mv

https://github.com/jarun/advcpmv