Functions: adds httpie helpers for xsrf
The functions are wrappers around http and https and set the
x-xsrf-token header based on the value of the Xsrf-token cookie in the
session file.

Currently, the cookie name and header name are hard coded. I may look to
specify those in an argument at some point although for now it's
probably fine.