Starts work on ActOnLast
This is a script that provides context menus (via rofi) based on a
previous action.

It takes an optional --first argument that, when present, causes the
script to select the first option on the list automatically.

It currently has two actions that it knows how to follow:

* using my password manager script to select a password
    - In this situation it will get a otp for the same password
* After taking a screenshot
    - open the screenshot
    - Add a shadow to the screenshot
    - Perform ocr on the screenshot

I will extend this to do more things in time.