Work on openapi completion
Currently still a work in progress.

the bin/.bin/openapi file will be a bash script for interfacing with an
openapi file. Currently it does things like listing requests, servers,
and viewing a request in json format.

The function added to fzf.zsh adds a zsh widget so that I can press a
keybinding (in my case ^o) to complete the command. It uses the above
script and fzf to make the selection. Currently supports http / https /
curl

The widget looks for an environment variable called OPENAPI which should
point to the openapi file. This is set in zshrc if a project is set.